requirements.txt
aiofiles==0.4.0
aiomysql==0.0.20
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
peewee==2.10.2
peewee-async==0.5.12
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
six==1.12.0
tornado==6.0.3
WTForms==2.2.1
WTForms-JSON==0.3.3
wtforms-tornado==0.0.2
